edit&show view

+++ resources/views/tasks/show.blade.php
@@ -1,28 +1,30 @@
 @extends('layouts.app')
 
 @section('content')
-
-    <h1>id = {{ $task->id }} のタスク詳細ページ</h1>
+    @if (Auth::check())
+        <h1>id = {{ $task->id }} のタスク詳細ページ</h1>
     
-     <table class="table table-bordered">
-        <tr>
-            <th>id</th>
-            <td>{{ $task->id }}</td>
-        </tr>
-        <tr>
-            <th>ステータス</th>
-            <td>{{ $task->status }}</td>
-        </tr>
-        <tr>
-            <th>タスク</th>
-            <td>{{ $task->content }}</td>
-        </tr>
-    </table>
+         <table class="table table-bordered">
+            <tr>
+                <th>id</th>
+                <td>{{ $task->id }}</td>
+            </tr>
+            <tr>
+                <th>ステータス</th>
+                <td>{{ $task->status }}</td>
+            </tr>
+            <tr>
+                <th>タスク</th>
+                <td>{{ $task->content }}</td>
+            </tr>
+        </table>
     
-    {!! link_to_route('tasks.edit', 'このタスク編集', ['id' => $task->id], ['class' => 'btn btn-default']) !!}
+        {!! link_to_route('tasks.edit', 'このタスク編集', ['id' => $task->id], ['class' => 'btn btn-default']) !!}
    
-    {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
-        {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
-    {!! Form::close() !!}
-    
+        {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
+            {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
+        {!! Form::close() !!}
+    @else
+        return redirect('/');
+    @endif
 @endsection
